+++ backend/canisters/group/impl/src/jobs/make_pending_payments.rs
@@ -2,12 +2,12 @@ use crate::{mutate_state, read_state, RuntimeState};
 use candid::Principal;
 use group_community_common::{PaymentRecipient, PendingPayment, PendingPaymentReason};
 use ic_cdk_timers::TimerId;
-use ic_ledger_types::BlockIndex;
 use icrc_ledger_types::icrc1::transfer::{Memo, TransferArg};
+use ledger_utils::icrc1::make_transfer;
 use std::cell::Cell;
 use std::time::Duration;
-use tracing::{error, trace};
-use types::{CanisterId, TimestampNanos};
+use tracing::trace;
+use types::TimestampNanos;
 use utils::consts::{MEMO_GROUP_IMPORT_INTO_COMMUNITY, MEMO_JOINING_FEE, SNS_GOVERNANCE_CANISTER_ID};
 
 thread_local! {
@@ -53,7 +53,7 @@ async fn process_payment(pending_payment: PendingPayment, now_nanos: TimestampNa
         amount: pending_payment.amount.into(),
     };
 
-    match make_payment(pending_payment.ledger_canister, args).await {
+    match make_transfer(pending_payment.ledger_canister, &args, true).await {
         Ok(_) => {
             if matches!(pending_payment.reason, PendingPaymentReason::AccessGate) {
                 if let PaymentRecipient::Member(user_id) = pending_payment.recipient {
@@ -66,28 +66,13 @@ async fn process_payment(pending_payment: PendingPayment, now_nanos: TimestampNa
                 }
             }
         }
-        Err(retry) => {
+        Err((_, retry)) => {
             if retry {
                 mutate_state(|state| {
                     state.data.pending_payments_queue.push(pending_payment);
                     start_job_if_required(state);
                 });
             }
-        }
-    }
-}
-
-// Error response contains a boolean stating if the transfer should be retried
-async fn make_payment(ledger_canister: CanisterId, transfer_args: TransferArg) -> Result<BlockIndex, bool> {
-    match icrc_ledger_canister_c2c_client::icrc1_transfer(ledger_canister, &transfer_args).await {
-        Ok(Ok(block_index)) => Ok(block_index.0.try_into().unwrap()),
-        Ok(Err(transfer_error)) => {
-            error!(?transfer_error, ?transfer_args, "Transfer failed");
-            Err(false)
-        }
-        Err(error) => {
-            error!(?error, ?transfer_args, "Transfer failed");
-            Err(true)
         }
     }
 }