+++ backend/libraries/ledger_utils/src/icrc1.rs
@@ -1,4 +1,6 @@
+use icrc_ledger_types::icrc1::transfer::TransferError;
 use icrc_ledger_types::icrc1::{account::Account, transfer::TransferArg};
+use tracing::error;
 use types::{
     icrc1::{CompletedCryptoTransaction, FailedCryptoTransaction, PendingCryptoTransaction},
     CanisterId,
@@ -7,6 +9,7 @@ use types::{
 pub async fn process_transaction(
     transaction: PendingCryptoTransaction,
     sender: CanisterId,
+    retry_if_bad_fee: bool,
 ) -> Result<CompletedCryptoTransaction, FailedCryptoTransaction> {
     let from = Account::from(sender);
 
@@ -19,8 +22,8 @@ pub async fn process_transaction(
         amount: transaction.amount.into(),
     };
 
-    match icrc_ledger_canister_c2c_client::icrc1_transfer(transaction.ledger, &args).await {
-        Ok(Ok(block_index)) => Ok(CompletedCryptoTransaction {
+    match make_transfer(transaction.ledger, &args, retry_if_bad_fee).await {
+        Ok(block_index) => Ok(CompletedCryptoTransaction {
             ledger: transaction.ledger,
             token: transaction.token.clone(),
             amount: transaction.amount,
@@ -29,16 +32,9 @@ pub async fn process_transaction(
             to: transaction.to.into(),
             memo: transaction.memo.clone(),
             created: transaction.created,
-            block_index: block_index.0.try_into().unwrap(),
+            block_index,
         }),
-        Ok(Err(transfer_error)) => {
-            let error_message = format!("Transfer failed. {transfer_error:?}");
-            Err(error_message)
-        }
-        Err((code, msg)) => {
-            let error_message = format!("Transfer failed. {code:?}: {msg}");
-            Err(error_message)
-        }
+        Err((error_message, _)) => Err(error_message),
     }
     .map_err(|error| FailedCryptoTransaction {
         ledger: transaction.ledger,
@@ -51,4 +47,48 @@ pub async fn process_transaction(
         created: transaction.created,
         error_message: error,
     })
+}
+
+// Error response contains the error message and a boolean stating if the transfer should be retried
+pub async fn make_transfer(
+    ledger_canister_id: CanisterId,
+    args: &TransferArg,
+    retry_if_bad_fee: bool,
+) -> Result<u64, (String, bool)> {
+    let mut response = icrc_ledger_canister_c2c_client::icrc1_transfer(ledger_canister_id, args).await;
+
+    if retry_if_bad_fee {
+        // If the ledger returns an error saying the fee is too high, reduce the fee and try again
+        if let Ok(Err(TransferError::BadFee { expected_fee })) = &response {
+            if let Some(fee) = args.fee.clone() {
+                let expected_fee = expected_fee.clone();
+                let mut updated_args = args.clone();
+                updated_args.fee = Some(expected_fee.clone());
+
+                if fee > expected_fee {
+                    let diff = fee - expected_fee;
+                    updated_args.amount += diff;
+                } else {
+                    let diff = expected_fee - fee;
+                    if updated_args.amount < diff {
+                        return Err(("Transfer amount too low to cover fee".to_string(), false));
+                    }
+                    updated_args.amount -= diff;
+                }
+                response = icrc_ledger_canister_c2c_client::icrc1_transfer(ledger_canister_id, &updated_args).await;
+            }
+        }
+    }
+
+    match response {
+        Ok(Ok(block_index)) => Ok(block_index.0.try_into().unwrap()),
+        Ok(Err(transfer_error)) => {
+            error!(?transfer_error, ?args, "Transfer failed");
+            Err((format!("Transfer failed. {transfer_error:?}"), false))
+        }
+        Err((code, msg)) => {
+            error!(?code, ?msg, ?args, "Transfer failed");
+            Err((format!("Transfer failed. {code:?}: {msg}"), true))
+        }
+    }
 }